requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
python-dotenv
passlib[bcrypt]
pytest
jose